requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15.2
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
h11==0.13.0
idna==3.4
imageio==2.21.3
networkx==2.8.6
numpy==1.23.3
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
pycodestyle==2.9.1
pyparsing==3.0.9
PySocks==1.7.1
pytz==2022.2.1
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.9.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tifffile==2022.8.12
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
